routes: migrate sandbox + oauth + image-generation domains (11 routes)
Move to:
- src/server/routes/sandbox.ts (6): sandbox-pool, worktree-pool, sandbox-status,
  sandbox-image/build, sandbox/host-tokens, connection-info
- src/server/routes/oauth.ts (4): status, flow-status, start, complete
- src/server/routes/image-generation.ts (1): generate

Routes registered in dispatcher; legacy handleApiRoute branches removed.
Drops now-unused imports from server.ts.

Net: server.ts -203 lines (7764 → 7561). No behaviour change.
